--- colors/everforest.lua
@@ -1,1 +1,0 @@
-require("everforest").load()
