+++ colors/forestflower.lua
@@ -1,0 +1,1 @@
+require("forestflower").load()
